Separação entre componente Post e componente Feed
