+++ spotted/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
-import { PostComponent } from './components/layout/post/post.component';
+import { PrincipalComponent } from './components/layout/principal/principal.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: 'home', pathMatch: 'full' },
-  { path: "home", component: PostComponent }
+  { path: "home", component: PrincipalComponent }
 ];
