+++ spotted/src/app/components/layout/feed/feed.component.html
@@ -1,10 +1,5 @@
 <article class="feed-container">
   @for (post of posts; track $index) {
-    <div *ngIf="post.valido">
-
-
-    </div>
- }
+    <app-post [post]="post"></app-post>
+  {
 </article>
-
-
