+++ spotted/src/app/components/layout/post/post.component.html
@@ -1,23 +1,16 @@
-<article class="feed-container">
-  @for (post of posts; track $index) {
-    <div *ngIf="post.valido">
-      <div class="card post" style="width: 38rem;">
-        <div class="card-header post">
-
-            <div class="post-container">
-              <img class="animal-pic" [src]="post.imagem" alt="img">
-          <p><strong>{{ post.imagemNome }} Anônimo disse:</strong> </p>
-            </div>
-       </div>
-        <div class="card-body">
-          <p><strong>{{ post.conteudo }}</strong></p>
-          <blockquote class="blockquote mb-0">
-            <footer class="blockquote-footer">{{ post.data | date:'medium'  }}</footer>
-          </blockquote>
-        </div>
+<div *ngIf="post.valido">
+  <div class="card post" style="width: 38rem;">
+    <div class="card-header post">
+      <div class="post-container">
+        <img class="animal-pic" [src]="post.imagem" alt="img">
+        <p><strong>{{ post.imagemNome }} Anônimo disse:</strong> </p>
       </div>
     </div>
- }
-</article>
-
-
+    <div class="card-body">
+      <p><strong>{{ post.conteudo }}</strong></p>
+      <blockquote class="blockquote mb-0">
+        <footer class="blockquote-footer">{{ post.data | date:'medium'  }}</footer>
+      </blockquote>
+    </div>
+  </div>
+</div>
